Remove schema and credential imports from devsetup_agent
